Roast chicken update
Added sword and damage
Added staple gun with damage
changed teleport (not finished yet)
Added new material (WIP)
Xiaolin Showdown

+++ PaperVR_V2/Assets/Script/PlayerStats.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using UnityEngine;
 
 public class PlayerStats : MonoBehaviour {
-    public int maxHealth = 6;
-    public int currenthealth;
-    public int damage = 1;
+    public float maxHealth = 6;
+    public float currenthealth;
+    public float damage = 1;
 
 	// Use this for initialization
 	void Start ()
